jobs_management/requirements.txt
Django==4.2.19
pip==25.0.1
djangorestframework==3.15.2